Fix asteroid explosion disappears on next level
The problem was that when the last asteroid was destroyed that
asteroid was considered dead before the explosion finished and
when it died all the current asteroids were deleted to prepare
for the new level which meant the last asteroid was deleted
before the explosion finsished.

Solve this by making asteroid.IsAlive() not return true before
the asteroid explosion is finished.