Fix destroyed asteroid does earth damage
The problem was caused by the previous change which changed
so that an asteroid is considered alive while it is exploding.

The reason this caused a problem is because of how it was
determined if an asteroid did earth damage.

 1. The update function first checked if the asteroid was alive.
 2. If it was then it updated the asteroid.
 3. Then it again checked if the asteroid was alive.
 4. If it was not alive the asteroid was considered to have died
    because it passed the player and therefore it should cause
	earth damage.

The reason this became a problem after the previous change was
because a destroyed asteroid would in step 2 update the explosion
and when the explosion finished then step 3 would return that
the asteroid now is dead.

This commit adds another check to make sure the asteroid died
without being destroyed.